Mac AVFoundation Video Capture: float frame rate
Floating point frame rate is supported in Chromium code now and is
also provided by AVFoundation, eliminate the unnecessary cast to int.

BUG=380463

Review URL: https://codereview.chromium.org/365953003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281260 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/video/capture/mac/video_capture_device_avfoundation_mac.mm
@@ -75,7 +75,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
            [format videoSupportedFrameRateRanges]) {
       media::VideoCaptureFormat format(
           gfx::Size(dimensions.width, dimensions.height),
-          static_cast<int>(frameRate.maxFrameRate),
+          frameRate.maxFrameRate,
           pixelFormat);
       formats->push_back(format);
       DVLOG(2) << name.name() << " resolution: "
